+++ Day17/Day17/Tools/Math/GroupRange.cs
@@ -25,7 +25,7 @@ namespace Day17.Tools.Math
             return _intervals.Any( x => x.Includes( number ) );
         }
 
-        public int GetNotIncludedValue()
+        public long GetNotIncludedValue()
         {
             if ( IsUnitary )
             {
