+++ Day17/Day17/Tools/Math/IPlaceable.cs
@@ -2,8 +2,8 @@ namespace Day17.Tools.Math
 {
     public interface IPlaceable
     {
-        public int CoordinateRow { get; set; }
+        public long CoordinateRow { get; set; }
 
-        public int CoordinateColumn { get; set;}
+        public long CoordinateColumn { get; set;}
     }
 }
